Reviewer1
First off great code writing. The pros’ of codes code are in the simple layout, for example starting off with the win1 and tab1 (not really sure why it ‘var win1’ is tabbed), then tacking on the ‘var Teas’ followed by ‘function getVerdict’ (line29). Made it very easy for myself to read down through it. I have also found that it makes it easier to compile and understand. Another pro of your code is that from what I have proofed is that your syntax is good.

Good use of notes, a couple more of descriptions (i.e. //Designing the compass app) maybe helpful to also input. A few critics would have to be in the way you didn’t utilize a creative color theme, or different corresponding msg to pop up (second character of the color). In designing your compass app, copied from the book looks spot on (chp.7 pg163. I hope you were fortune enough to get your emulator running, in designing mine it wasn’t so much the case.